refactor(domain/notification): unifies notification related entities into services folder
