--- libs/atqr/domain/src/lib/notifications/index.ts
@@ -1,1 +1,0 @@
-export * from './templateDataType';
