+++ libs/atqr/domain/src/lib/services/index.ts
@@ -1,2 +1,2 @@
 export * from './deadlineMonitor.service';
-export * from './notification.service';
+export * from './notification/';
