+++ libs/atqr/domain/src/lib/services/notification/index.ts
@@ -1,0 +1,2 @@
+export * from './notification.service';
+export * from './notification.types';
