--- libs/atqr/domain/src/lib/services/notificationsEnums.ts
@@ -1,29 +1,0 @@
-export enum NotificationChannel {
-  Email = 'Email',
-  SMS = 'SMS',
-  Push = 'Push',
-}
-
-export enum MessageTemplateId {
-  ChallengeCreated = 'ChallengeCreated',
-  ChallengeCanceled = 'ChallengeCanceled',
-  ChallengeOngoing = 'ChallengeOngoing',
-  ChallengeCompleted = 'ChallengeCompleted',
-  ChallengeFailed = 'ChallengeFailed',
-  YourOfficiationHaveBeenRequested = 'YourOfficiationHaveBeenRequested',
-  YourOfficiationRequestHaveBeenAccepted = 'YourOfficiationRequestHaveBeenAccepted',
-  YourOfficiationRequestHaveBeenRejected = 'YourOfficiationRequestHaveBeenRejected',
-  YourOfficiationRequestHaveBeenIgnored = 'YourOfficiationRequestHaveBeenIgnored',
-  YouHaveBeenChallenged = 'YouHaveBeenChallenged',
-  YourContenderAcceptedTheChallenge = 'YourContenderAcceptedTheChallenge',
-  YourContenderRejectedTheChallenge = 'YourContenderRejectedTheChallenge ',
-  YourContenderIgnored = 'YourContenderIgnored',
-  YourContenderRemovedYouFromJudgeRole = 'YourContenderRemovedYouFromJudgeRole',
-  YourContenderCanceledTheChallenge = 'YourContenderCanceledTheChallenge',
-  YourContenderWereTheySuccessful = 'YourContenderWereTheySuccessful',
-  TimeToOfficiateArrived = 'TimeToOfficiateArrived',
-  YourJudgeDidNotAnswered = 'YourJudgeDidNotAnswered',
-}
-
-
-
